backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]
SQLAlchemy
python-multipart
weaviate-client
pypdf
docx2txt
google-generativeai
sentence-transformers
numpy
fastapi-cors
scikit-learn